running data collections within server analytics
