+++ scripts/dataGenerator.js
@@ -1080,13 +1080,19 @@ db.initialize(function () {
                 process.exit(1);
             }
 
-            db.resetAllSettings(function (err, res) {
+            db.removeAnalytics(function (err, res) {
                 if (err) {
                     process.exit(1);
                 }
 
-                calculateTotalNumberOfQuestions();
-                createAdmins();
+                db.resetAllSettings(function (err, res) {
+                    if (err) {
+                        process.exit(1);
+                    }
+    
+                    calculateTotalNumberOfQuestions();
+                    createAdmins();
+                });
             });
         });
     });
